Few bug fixes + build ee offline

+++ src/main/java/timeseriesweka/classifiers/distance_based/distance_measures/DistanceMeasureParameterSpaces.java
@@ -55,7 +55,7 @@ public class DistanceMeasureParameterSpaces {
     public static ParameterSpace buildWdtwParameterSpace() {
         double[] gValues = new double[100];
         for(int i = 0; i < gValues.length; i++) {
-            gValues[i] = i / gValues.length;
+            gValues[i] = (double) i / gValues.length;
         }
         ParameterSpace parameterSpace = new ParameterSpace();
         parameterSpace.addParameter(Wdtw.DISTANCE_MEASURE_KEY, Wdtw.NAME);
